feat(QueryGeneration): refactor conversation handling and introduce QueryGenerationRule
