+++ core/src/commonMain/kotlin/dev/gabrielolv/kaia/core/HandoffManager.kt
@@ -3,7 +3,6 @@ package dev.gabrielolv.kaia.core
 import dev.gabrielolv.kaia.core.model.*
 import dev.gabrielolv.kaia.core.tenant.TenantContext
 import dev.gabrielolv.kaia.core.tenant.withTenantContext
-import dev.gabrielolv.kaia.core.tools.ToolExecutionFailedException
 import dev.gabrielolv.kaia.core.tools.ToolManager
 import dev.gabrielolv.kaia.llm.LLMMessage
 import dev.gabrielolv.kaia.utils.nextThreadId
@@ -19,7 +18,7 @@ import kotlinx.serialization.json.Json
 data class AgentTaskContext(
     val originalUserRequest: String,
     val currentTask: String,
-    val reasonForTask: String? // Added: Context from Director
+    val reasonForTask: String?
 )
 
 class HandoffManager(
@@ -34,13 +33,9 @@ class HandoffManager(
      * Start a new conversation. The initial agent is now the planner.
      */
     suspend fun startConversation(
-        conversationId: String = nextThreadId,
-        initialMessageContent: String
+        conversationId: String = nextThreadId
     ): String {
-        val conversation = Conversation(
-            id = conversationId,
-            originalUserRequest = initialMessageContent
-        )
+        val conversation = Conversation(id = conversationId)
         lock.withLock { conversations[conversationId] = conversation }
         return conversationId
     }
@@ -61,11 +56,6 @@ class HandoffManager(
         val conversation = lock.withLock { conversations[conversationId] }
             ?: return null // Or throw exception
 
-        // Ensure original request is set if this is the very first user message
-        if (conversation.messages.none { it is LLMMessage.UserMessage }) {
-            conversation.originalUserRequest = message.content
-        }
-
         return channelFlow {
             withTenantContext(tenantContext) {
                 val userMessage = LLMMessage.UserMessage(content = message.content)
@@ -75,6 +65,7 @@ class HandoffManager(
                 try {
                     manageStepByStepExecution(
                         conversation,
+                         message,
                         directorAgentId,
                         message,
                         this@channelFlow
@@ -87,7 +78,7 @@ class HandoffManager(
                     conversation.append(errorMsg)
                 } finally {
                     val lastExecuted = conversation.executedSteps.lastOrNull()
-                    if (lastExecuted?.status == StepStatus.COMPLETED && conversation.executedSteps.size > 0) {
+                    if (lastExecuted?.status == StepStatus.COMPLETED && conversation.executedSteps.isNotEmpty()) {
                         send(LLMMessage.SystemMessage("Processing complete."))
                     } else if (conversation.executedSteps.any { it.status == StepStatus.FAILED }) {
                         send(LLMMessage.SystemMessage("Processing finished with errors."))
@@ -102,6 +93,7 @@ class HandoffManager(
 
     private suspend fun manageStepByStepExecution(
         conversation: Conversation,
+        message: LLMMessage.UserMessage,
         directorAgentId: String,
         triggerMessage: LLMMessage.UserMessage,
         scope: ProducerScope<LLMMessage>
@@ -178,14 +170,14 @@ class HandoffManager(
                 break
             }
 
-            emitAndStore(LLMMessage.SystemMessage("Director decision: ${directorOutput!!.reasoningTrace}"))
+            emitAndStore(LLMMessage.SystemMessage("Director decision: ${directorOutput.reasoningTrace}"))
 
-            if (directorOutput!!.isComplete) {
+            if (directorOutput.isComplete) {
                 emitAndStore(LLMMessage.SystemMessage("Director indicates task is complete."))
                 break
             }
 
-            val nextStepInfo = directorOutput!!.nextStep
+            val nextStepInfo = directorOutput.nextStep
             if (nextStepInfo == null) {
                 emitAndStore(LLMMessage.SystemMessage("Director indicates task is not complete, but provided no next step. Halting."))
                 break
@@ -224,7 +216,7 @@ class HandoffManager(
             try {
                 // Prepare context for the agent (similar to before)
                 val agentContext = AgentTaskContext(
-                    originalUserRequest = conversation.originalUserRequest,
+                    originalUserRequest = message.content,
                     currentTask = nextStepInfo.action,
                     reasonForTask = nextStepInfo.reason
                 )
@@ -286,25 +278,23 @@ class HandoffManager(
                                                     )
                                                 )
                                             }.onLeft { error ->
-                                                emitAndStore(LLMMessage.SystemMessage("Tool execution failed for ${toolCall.name}: $error"))
-                                                executedStepRecord.error = "Tool failed: ${error}"
+                                                val toolErrorMsg =
+                                                    LLMMessage.SystemMessage("Tool execution failed for ${toolCall.name}: $error")
+                                                emitAndStore(toolErrorMsg)
+                                                executedStepRecord.messages.add(toolErrorMsg)
+                                                executedStepRecord.error = "Tool exception: $error"
+
+                                                executedStepRecord.error = "Tool failed: $error"
+                                                conversation.append(
+                                                    LLMMessage.ToolResponseMessage(
+                                                        toolCall.id,
+                                                        "Execution Error: $error"
+                                                    )
+                                                )
                                                 stepFailed = true // Mark step as failed if tool fails
                                             }
 
 
-                                    } catch (toolEx: ToolExecutionFailedException) {
-                                        val toolErrorMsg =
-                                            LLMMessage.SystemMessage("Tool execution exception for ${toolCall.name}: ${toolEx.message}")
-                                        emitAndStore(toolErrorMsg)
-                                        executedStepRecord.messages.add(toolErrorMsg)
-                                        executedStepRecord.error = "Tool exception: ${toolEx.message}"
-                                        conversation.append(
-                                            LLMMessage.ToolResponseMessage(
-                                                toolCall.id,
-                                                "Execution Error: ${toolEx.message}"
-                                            )
-                                        )
-                                        stepFailed = true
                                     } catch (e: Exception) {
                                         val toolErrorMsg =
                                             LLMMessage.SystemMessage("Unexpected error during tool execution for ${toolCall.name}: ${e.message}")
